requirements.txt
anyio==4.12.1
certifi==2026.4.22
exceptiongroup==1.3.1
greenlet==3.2.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.15
playwright==1.59.0
pyee==13.0.1
python-dotenv==1.2.1
python-telegram-bot==22.5
typing_extensions==4.15.0
